Feat/#114: 툴 상세 정보 조회 API 공통으로 이동
