+++ src/pages/toolDetail/apis/api.ts
@@ -1,23 +1,11 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getCoreFeature, getDetail, getPlan, getRelatedTool } from './queries';
+import { getCoreFeature, getPlan, getRelatedTool } from './queries';
 
 export const DETAIL_QUERY_KEY = {
-  TOOL_DETAIL: (toolID: number) => ['tooldetail', toolID],
   CORE_FEATURES: (coreID: number) => ['corefeature', coreID],
   TOOL_PLAN: (planID: number) => ['toolplan', planID],
   RELATED_TOOLS: (toolID: number) => ['relatedtool', toolID],
-};
-
-// 툴 상세 정보 가져오기
-export const useToolData = (toolId: number) => {
-  return useQuery({
-    queryKey: DETAIL_QUERY_KEY.TOOL_DETAIL(toolId),
-    queryFn: () => getDetail(toolId),
-    staleTime: 1000 * 60 * 60,
-    gcTime: 1000 * 60 * 60 * 24,
-    enabled: !!toolId,
-  });
 };
 
 // 핵심 기능 조회하기
